requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
cycler==0.11.0
daphne==3.0.2
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
fonttools==4.28.5
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kiwisolver==1.3.2
matplotlib==3.5.1
msgpack==1.0.3
numpy==1.22.0
orjson==3.6.5
packaging==21.3
Pillow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
uvicorn==0.16.0
zope.interface==5.4.0
